requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
pydantic==1.10.9
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
httpx==0.24.1
python-multipart
email-validator
